fix existing issues
